Move non-legacy static sites code out of legacy
